categorien toegevoegd

+++ php/deletefolderorfile.php
@@ -10,17 +10,40 @@
         
         //See if a file or folder needs to be removed
         if (strtolower($type) == "folder") {
-            
+
+            //get pathname and foldername in 1var
             $pathname = $dirPath . "/" . $fileName;
-            if (rmdir($pathname)) {
-                echo "<p class='alert alert-success' role='alert'>Folder '".$fileName."' has successfully been removed</p>";
-                die(); 
+
+            //Check if folder is empty
+            if (!$testing = glob($pathname . "/*")) {
+
+                //Remove folder
+                if (rmdir($pathname)) {
+                    echo "<p class='alert alert-success' role='alert'>Folder '".$fileName."' has successfully been removed</p>";
+                    die(); 
+    
+                } else {
+                    echo "<p class='alert alert-danger' role='alert'>Failed to remove folder '".$fileName."'</p>";
+                    die(); 
+                }//if rmdir
+              
             } else {
-                echo "<p class='alert alert-danger' role='alert'>Failed to remove folder</p>";
-                die(); 
-            }
+                echo "<p class='alert alert-danger' role='alert'>Failed to remove folder '".$fileName."'. The folder is not empty!</p>";
+                die();
+            }//if folder == empty
             
         } else if (strtolower($type) == "file") {
+
+            //Remove file
+            $fileRemove = $dirPath . "/" . $fileName;
+            if (unlink($fileRemove)) {
+                echo "<p class='alert alert-success' role='alert'>File '".$fileName."' has successfully been removed</p>";
+                die(); 
+
+            } else {
+                echo "<p class='alert alert-danger' role='alert'>Failed to remove folder '".$fileName."'</p>";
+                die();
+            }//if unlink
             
         } else {
             echo "<p class='alert alert-danger' role='alert'>Incorrect file/folder type</p>";
